test: refactored erc165 singleton deployment
