+++ test/extensions/erc165.test.js
@@ -2,14 +2,12 @@ const { BigNumber } = require("ethers");
 const { ethers } = require("hardhat");
 const chai = require("chai");
 const { solidity } = require("ethereum-waffle");
+const { erc165SingletonAddress } = require("../utils/constants");
+const { deployERC165Singleton } = require("../utils/utils");
 chai.use(solidity);
 const { expect, assert } = chai;
 
 describe("ERC165Logic", function () {
-    const singletonFactoryDeploymentTx = "0xf9016c8085174876e8008303c4d88080b90154608060405234801561001057600080fd5b50610134806100206000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c80634af63f0214602d575b600080fd5b60cf60048036036040811015604157600080fd5b810190602081018135640100000000811115605b57600080fd5b820183602082011115606c57600080fd5b80359060200191846001830284011164010000000083111715608d57600080fd5b91908080601f016020809104026020016040519081016040528093929190818152602001838380828437600092019190915250929550509135925060eb915050565b604080516001600160a01b039092168252519081900360200190f35b6000818351602085016000f5939250505056fea26469706673582212206b44f8a82cb6b156bfcc3dc6aadd6df4eefd204bc928a4397fd15dacf6d5320564736f6c634300060200331b83247000822470";
-    const singletonFactoryDeployer = "0xBb6e024b9cFFACB947A71991E386681B1Cd1477D";
-    const singletonFactoryAddress = "0xce0042B868300000d44A59004Da54A005ffdcf9f";
-    const erc165Address = "0x23A6e4d33CFF52F908f3Ed8f7E883D2A91A4918f";
     const interfaceId = "0x80ac58cd";
     const invalidInterface = "0xffffffff";
     let account;
@@ -19,18 +17,11 @@ describe("ERC165Logic", function () {
     before("deploy new", async function () {
         [account, account2] = await ethers.getSigners();
         
+        await deployERC165Singleton(account);
+
         const ERC165Logic = await ethers.getContractFactory("ERC165Logic");
-        const bytecode = (await ERC165Logic.getDeployTransaction()).data;
-
-        await account.sendTransaction({ to: singletonFactoryDeployer, value: ethers.utils.parseEther("1") });
-        await ethers.provider.sendTransaction(singletonFactoryDeploymentTx);
-
-        const Factory = new hre.ethers.Contract("0x0000000000000000000000000000000000000000", factoryABI, account);
-        const factory = await Factory.attach(singletonFactoryAddress);
-        await factory.deploy(bytecode, "0x0000000000000000000000000000000000000000000000000000000000000000", { gasLimit: "0x07A120" });
-
         const ERC165Caller = await ethers.getContractFactory("ERC165Caller");
-        erc165 = await ERC165Logic.attach(erc165Address);
+        erc165 = await ERC165Logic.attach(erc165SingletonAddress);
         erc165Caller = await ERC165Caller.deploy(erc165.address);
     })
